extended hal

+++ src/order/command_queue.c
@@ -271,7 +271,7 @@ static void send_queue_failed_response(uint_fast8_t cause, uint_fast8_t enqueued
 {
     uint_fast16_t slots_available;
     uint_fast16_t slots_taken;
-    uint_fast8_t *para_start = com_get_start_parameter();
+    uint8_t *para_start = com_get_start_parameter();
     if(read_pos <= write_pos)
     {
         slots_taken = write_pos - read_pos;
@@ -282,18 +282,18 @@ static void send_queue_failed_response(uint_fast8_t cause, uint_fast8_t enqueued
     }
     slots_available = MAX_QUEUE_ELEMENTS - slots_taken;
 
-    *(para_start + 0) = cause;
-    *(para_start + 1) = enqueued_commands;
+    *(para_start + 0) = (uint8_t)cause;
+    *(para_start + 1) = (uint8_t)enqueued_commands;
 
-    *(para_start + 2) = (slots_available >> 8) & 0xff;
-    *(para_start + 3) = slots_available  & 0xff;
+    *(para_start + 2) = (uint8_t)(slots_available >> 8) & 0xff;
+    *(para_start + 3) = (uint8_t)slots_available  & 0xff;
 
-    *(para_start + 4) = (slots_taken >> 8) & 0xff;
-    *(para_start + 5) = slots_taken & 0xff;
+    *(para_start + 4) = (uint8_t)(slots_taken >> 8) & 0xff;
+    *(para_start + 5) = (uint8_t)slots_taken & 0xff;
 
-    *(para_start + 6) = (finished_blocks >> 8) & 0xff;
-    *(para_start + 7) = finished_blocks & 0xff;
-    *(para_start + 8) = 0xff;
+    *(para_start + 6) = (uint8_t)(finished_blocks >> 8) & 0xff;
+    *(para_start + 7) = (uint8_t)finished_blocks & 0xff;
+    *(para_start + 8) = (uint8_t)0xff;
     com_send_order_specific_error_with_prefilled_parameter(9);
 }
 
@@ -301,7 +301,7 @@ static void send_queue_ok_response(void)
 {
     uint_fast16_t slots_available;
     uint_fast16_t slots_taken;
-    uint_fast8_t *para_start = com_get_start_parameter();
+    uint8_t *para_start = com_get_start_parameter();
     if(read_pos <= write_pos)
     {
         slots_taken = write_pos - read_pos;
@@ -312,14 +312,14 @@ static void send_queue_ok_response(void)
     }
     slots_available = MAX_QUEUE_ELEMENTS - slots_taken;
 
-    *(para_start + 0) = (slots_available >> 8) & 0xff;
-    *(para_start + 1) = slots_available  & 0xff;
+    *(para_start + 0) = (uint8_t)(slots_available >> 8) & 0xff;
+    *(para_start + 1) = (uint8_t)slots_available  & 0xff;
 
-    *(para_start + 2) = (slots_taken >> 8) & 0xff;
-    *(para_start + 3) = slots_taken & 0xff;
+    *(para_start + 2) = (uint8_t)(slots_taken >> 8) & 0xff;
+    *(para_start + 3) = (uint8_t)slots_taken & 0xff;
 
-    *(para_start + 2) = (finished_blocks >> 8) & 0xff;
-    *(para_start + 3) = finished_blocks & 0xff;
+    *(para_start + 2) = (uint8_t)(finished_blocks >> 8) & 0xff;
+    *(para_start + 3) = (uint8_t)finished_blocks & 0xff;
     com_send_ok_with_prefilled_parameter(6);
 }
